Refactor problem page
