+++ src/components/TestCaseSolver/index.tsx
@@ -1,0 +1,3 @@
+import TestCaseSolver from './TestCaseSolver';
+
+export default TestCaseSolver;
